breaking up row removal methods
